Fix: ia16-elf-gcc code wrongly assumed XMS driver preserves %bx
This was causing XMSinit() to compute a lower value for
XMSsave.length than needed --- gcc had copied SwapTransientSize
into %bx, and `lcall *%cs:XMSdriverAdress' clobbered %bl to 0.
The effect was that part of the command.com stack was not
swapped out to extended memory as it should be.

I also took the opportunity to tweak a few other `asm volatile
(...)' statements, partly to make sure that `asm' clobber list
does not overlap with the input registers (as required by the
GCC manual).

One memory corruption bug down, more to go...