Supplement: Match fmem* functions to common prototypes
Since ia16-elf-gcc's libi86 received support for fmemcpy, fmemmove
and fmemset, its prototypes are in conflict with the replacement
functions defined in the suppl directory. If libi86 implements all
of the support required for FreeCOM in the future, then we may not
need to include suppl, but until then adjust the suppl to match the
common prototypes.

Patch by @tkchia, and discussion at [fixes #79]