fixes #61 - Existence in question, properly report error for "IF EXIST file"
"IF EXIST file ECHO hi" followed by "IF EXIST file" would advance pp to the "ECHO hi" portion, but "IF EXIST file " followed by "IF EXIST file" (note the space at end) would correctly advance to "" and report error.  Don't increment pp as it is already '\0' terminated and points to a delimiter or end of string and both should work correctly.