mcb_walk.c: try querying UMB link state even if not on DOS 5+
Current lMS-DOS still reports its version as v4.00 but does fully
support a DOS-managed UMA. As of today lMS-DOS also supports the
SHELLHIGH= directive. Like FreeDOS's this is implemented using
the 21.4B80 extension. Although the internals differ, the outside
observable behaviour is the DOS executable loader allocates both
the environment and the process memory block into the UMA, if
there are free UMBs. However, unlike LH the UMB link state is not
enabled once the control flow is passed to the application. That
means an environment block in a UMB isn't reachable when walking
the MCBs without regard to proceeding past the first UMCB.

The MCB walker is used in FreeCOM's init to validate the
environment block passed to it upon its startup. Reference:
https://github.com/FDOS/freecom/blob/cede733377c454cb93e9038d476582f0d1fcbf0a/shell/init.c#L238

If this check fails, then the shell will prompt for the position
of its executable as follows. This patch makes the MCB walker
enable the UMB link even if the reported DOS version is below 5,
allowing it to find the environment block's UMCB and avoiding
the prompt.

Failed to load the strings resource into memory, the location
pointed to in %COMSPEC% seems to be invalid. Please specify another
location of FreeCOM to try to load the strings from, e.g.:
C:\COMMAND.COM
or just hit enter to cancel to load the strings.