set CY before calling 21.71 or 21.73 DOS functions (LFN, FAT32)
Between the current lDOS kernel (which doesn't support 21.73),
SRDISK, and dosemu2 revision 6a2f4f527 the return from 21.7303
somehow is with both Carry Flag and AX preserved. This makes the
DIR command think that the call is supported when it isn't.

The proper check is to call with CY set before the call.